services/spark-processor/requirements.txt
pyspark==3.5.0
apache-sedona
h3
confluent-kafka
